[dg mcp] move configure to mcp group
